requirement.txt
Django==4.0.4
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
djangorestframework==3.13.1
idna==3.3
pip==22.0.4
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
setuptools==60.10.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
